extracted functions from evm adaptors
